feat: add document management (wip)
