+++ app/web_ui/src/routes/(app)/documents/[project_id]/documents/+page.ts
